Moved plotly-backed plotting methods to their own files
